드래곤 교배 로직 에러 해결

+++ contracts/dragonNFT/interface/IDragonNFT.sol
@@ -5,7 +5,7 @@ import "../library/DragonNFTLib.sol";
 
 interface IDragonNFT {
     function mintNewDragon(address requester, uint256[] memory _randomWords) external;
-    function createDragon(address _to, DragonNFTLib.Gender _gender, DragonNFTLib.Rarity _rarity, DragonNFTLib.Species _specie, uint64 _damage, uint32 _xpPerSec) external returns(uint256 tokenId);
+    function createDragon(address _to, DragonNFTLib.Gender _gender, DragonNFTLib.Rarity _rarity, DragonNFTLib.Species _specie, uint16 _damage, uint8 _xpPerSec) external returns(uint256 tokenId);
     function addExperience(uint256 tokenId) external;
     function setXpToLevelUp(uint8 level, uint128 xpRequired) external;
     function getDragonInfo(uint256 tokenId) external view returns(DragonNFTLib.Dragon memory dragonInfo);
